fix eslint issues
