+++ frontend/src/app/App.tsx
@@ -6,8 +6,8 @@ import ReactDOM from "react-dom/client";
 import "./index.css";
 import { initReactI18next } from "react-i18next";
 import auth from "@/shared/locales/en/auth.json";
-import product from "@/shared/locales/en/product.json";
 import principal from "@/shared/locales/en/principal.json";
+import product from "@/shared/locales/en/product.json";
 import { routeTree } from "@/routeTree.gen";
 
 const queryClient = new QueryClient();
