+++ frontend/src/app/layout/layout.tsx
@@ -1,7 +1,15 @@
-import { Outlet } from "@tanstack/react-router";
+import { Link, Outlet } from "@tanstack/react-router";
+import { Plus } from "lucide-react";
+import { Suspense } from "react";
+import { useTranslation } from "react-i18next";
 import { PrincipalSidebarFooter } from "@/entities/principal";
+import { ProductsSidebar, ProductsSidebarSkeleton } from "@/entities/product";
 import { AppSidebar } from "@/shared/components/app-sidebar.tsx";
 import {
+  SidebarGroup,
+  SidebarGroupAction,
+  SidebarGroupContent,
+  SidebarGroupLabel,
   SidebarInset,
   SidebarProvider,
   SidebarTrigger,
@@ -10,7 +18,10 @@ import {
 export const Layout = () => {
   return (
     <SidebarProvider>
-      <AppSidebar principalSlot={<PrincipalSidebarFooter />} />
+      <AppSidebar
+        principalSlot={<PrincipalSidebarFooter />}
+        productsSlot={<SidebarProducts />}
+      />
       <SidebarInset>
         <header className="flex h-16 shrink-0 items-center gap-2 border-b px-4">
           <SidebarTrigger />
@@ -22,3 +33,25 @@ export const Layout = () => {
     </SidebarProvider>
   );
 };
+
+function SidebarProducts() {
+  const { t } = useTranslation("product");
+  return (
+    <SidebarGroup>
+      <SidebarGroupLabel asChild>
+        <Link to="/products">{t("sidebar.title")}</Link>
+      </SidebarGroupLabel>
+      <SidebarGroupAction title={t("sidebar.actions.create.label")} asChild>
+        <Link to="/products/create">
+          <Plus />
+          <span className="sr-only">{t("sidebar.actions.create.label")}</span>
+        </Link>
+      </SidebarGroupAction>
+      <SidebarGroupContent>
+        <Suspense fallback={<ProductsSidebarSkeleton />}>
+          <ProductsSidebar />
+        </Suspense>
+      </SidebarGroupContent>
+    </SidebarGroup>
+  );
+}
