+++ frontend/src/app/routes/_authenticated/products/$productId.issues.tsx
@@ -1,12 +1,11 @@
-import * as React from 'react'
-import { createFileRoute } from '@tanstack/react-router'
+import { createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute(
-  '/_authenticated/products/$productId/issues',
+  "/_authenticated/products/$productId/issues",
 )({
   component: RouteComponent,
-})
+});
 
 function RouteComponent() {
-  return 'Hello /_authenticated/products/$productId/issues!'
+  return "Hello /_authenticated/products/$productId/issues!";
 }
