+++ frontend/src/app/routes/_authenticated/products/$productId.tsx
@@ -1,10 +1,9 @@
-import * as React from 'react'
-import { createFileRoute } from '@tanstack/react-router'
+import { createFileRoute } from "@tanstack/react-router";
 
-export const Route = createFileRoute('/_authenticated/products/$productId')({
+export const Route = createFileRoute("/_authenticated/products/$productId")({
   component: RouteComponent,
-})
+});
 
 function RouteComponent() {
-  return 'Hello /_authenticated/products/$productId!'
+  return "Hello /_authenticated/products/$productId!";
 }
