+++ frontend/src/app/routes/_authenticated/products/create.tsx
@@ -1,10 +1,9 @@
-import * as React from 'react'
-import { createFileRoute } from '@tanstack/react-router'
+import { createFileRoute } from "@tanstack/react-router";
 
-export const Route = createFileRoute('/_authenticated/products/create')({
+export const Route = createFileRoute("/_authenticated/products/create")({
   component: RouteComponent,
-})
+});
 
 function RouteComponent() {
-  return 'Hello /_authenticated/products/create!'
+  return "Hello /_authenticated/products/create!";
 }
