+++ frontend/src/entities/user/api.ts
@@ -4,5 +4,5 @@ export const userSchema = z.object({
   id: z.number(),
   fullName: z.string(),
   username: z.string(),
-  role: z.enum(["ADMIN", "USER"])
-})
+  role: z.enum(["ADMIN", "USER"]),
+});
