+++ frontend/src/entities/user/index.ts
@@ -1,1 +1,1 @@
-export * from './api';
+export * from "./api";
